requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy[asyncio]
asyncpg
pydantic[email]
pydantic-settings
alembic
minio
aiofiles
celery[redis]
redis
Pillow
PyPDF2
pdf2image
python-dateutil
tenacity
structlog
python-dotenv
httpx
PyJWT>=2.8.0

# Core Dependencies for Ensemble DEID Application

# Presidio - Microsoft's PII detection and anonymization framework
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# spaCy and transformers integration
spacy>=3.7.0
spacy-transformers>=1.3.0
spacy-huggingface-pipelines>=0.0.1
# Transformers - Hugging Face transformers for Stanford model
transformers>=4.30.0
torch>=2.0.0

# Additional dependencies (usually installed with above packages)
numpy>=1.24.0
requests>=2.31.0
openai
# Note: After installation, you need to download:
# 1. spaCy English model: python -m spacy download en_core_web_sm
# 2. Transformers will auto-download "StanfordAIMI/stanford-deidentifier-base" on first run

